well that was fast to clean up

+++ launch_service_c/src/launch_service_c.cpp
@@ -65,7 +65,7 @@ int main(int argc, char **argv)
         }
         // Create extra launch arguments
         std::vector<std::string> launch_args;
-        for (size_t i = child_flag_index + 2; i < argc; ++i)
+        for (int i = child_flag_index + 2; i < argc; ++i)
         {
             launch_args.push_back(argv[i]);
         }
